+++ src/main/java/Nat/Deadline.java
@@ -1,13 +1,13 @@
+package Nat;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task {
-    private LocalDateTime startDate;
+public class Deadline extends Task {
     private LocalDateTime dueDate;
 
-    public Event(String taskName, String startDate, String dueDate) {
+    public Deadline(String taskName, String dueDate) {
         super(taskName);
-        this.startDate = this.parseDateTime(startDate);
         this.dueDate = this.parseDateTime(dueDate);
     }
 
@@ -18,20 +18,19 @@ public class Event extends Task {
 
     @Override
     public String toSaveFormat() {
-        // Format: "T | 1 | Read a book | 5pm | 9pm"
+        // Format: "T | 1 | Read a book | 21 Mar 25"
         DateTimeFormatter format = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
-        return super.toSaveFormat() + " | " + this.startDate.format(format) + " | " + this.dueDate.format(format);
+        return super.toSaveFormat() + " | " + this.dueDate.format(format);
     }
 
     @Override
     public String getTaskType() {
-        return "E";
+        return "D";
     }
 
     @Override
     public String toString() {
         DateTimeFormatter outputFormat = DateTimeFormatter.ofPattern("MMM dd yyyy, h:mm a");
-        return "[E]" + super.toString() + " (from: " + this.startDate.format(outputFormat) + " to: "
-                + this.dueDate.format(outputFormat) + ")";
+        return "[D]" + super.toString() + " (by: " + this.dueDate.format(outputFormat) + ")";
     }
 }
