+++ src/main/java/Nat/Event.java
@@ -1,11 +1,15 @@
+package Nat;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task {
+public class Event extends Task {
+    private LocalDateTime startDate;
     private LocalDateTime dueDate;
 
-    public Deadline(String taskName, String dueDate) {
+    public Event(String taskName, String startDate, String dueDate) {
         super(taskName);
+        this.startDate = this.parseDateTime(startDate);
         this.dueDate = this.parseDateTime(dueDate);
     }
 
@@ -16,19 +20,20 @@ public class Deadline extends Task {
 
     @Override
     public String toSaveFormat() {
-        // Format: "T | 1 | Read a book | 21 Mar 25"
+        // Format: "T | 1 | Read a book | 5pm | 9pm"
         DateTimeFormatter format = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
-        return super.toSaveFormat() + " | " + this.dueDate.format(format);
+        return super.toSaveFormat() + " | " + this.startDate.format(format) + " | " + this.dueDate.format(format);
     }
 
     @Override
     public String getTaskType() {
-        return "D";
+        return "E";
     }
 
     @Override
     public String toString() {
         DateTimeFormatter outputFormat = DateTimeFormatter.ofPattern("MMM dd yyyy, h:mm a");
-        return "[D]" + super.toString() + " (by: " + this.dueDate.format(outputFormat) + ")";
+        return "[E]" + super.toString() + " (from: " + this.startDate.format(outputFormat) + " to: "
+                + this.dueDate.format(outputFormat) + ")";
     }
 }
