+++ src/main/java/Nat/Task.java
@@ -1,3 +1,5 @@
+package Nat;
+
 public class Task {
     protected String taskName;
     protected boolean isDone;
