+++ src/main/java/Nat/Ui.java
@@ -1,3 +1,5 @@
+package Nat;
+
 import java.util.ArrayList;
 
 public class Ui {
